Add value package for value object
